feat: using lerna for mono repo, redesign highlight code block
