+++ package.json
@@ -1,24 +1,18 @@
 {
   "name": "engineer",
   "private": true,
+  "workspaces": [
+    "packages/*"
+  ],
+  "version": "1.0.0",
+  "repository": "https://github.com/youngtailorshq/engineer",
+  "author": "Evan Nguyen <dacsang97@gmail.com>",
+  "license": "MIT",
   "scripts": {
-    "build": "gridsome build",
-    "dev": "gridsome develop",
-    "explore": "gridsome explore"
-  },
-  "dependencies": {
-    "@gridsome/vue-remark": "^0.2.1",
-    "gridsome": "^0.7.0",
-    "tailwindcss": "^1.4.6",
-    "vue-scrollactive": "^0.9.3"
+    "dev": "lerna exec --scope engineer yarn dev",
+    "build": "lerna exec --scope engineer yarn build"
   },
   "devDependencies": {
-    "gridsome-plugin-remark-shiki": "^0.3.1",
-    "postcss-import": "^12.0.1",
-    "postcss-import-url": "^5.1.0",
-    "postcss-preset-env": "^6.7.0",
-    "shiki": "^0.1.7",
-    "shiki-languages": "^0.1.6",
-    "unist-util-visit": "^2.0.2"
+    "lerna": "^3.22.0"
   }
 }
