+++ packages/engineer/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "engineer",
+  "version": "0.0.0",
+  "scripts": {
+    "build": "gridsome build",
+    "dev": "gridsome develop",
+    "explore": "gridsome explore"
+  },
+  "dependencies": {
+    "@gridsome/vue-remark": "^0.2.1",
+    "gridsome": "^0.7.0",
+    "vue-scrollactive": "^0.9.3",
+    "yt-shiki": "^0.0.1",
+    "ytailwind": "^0.0.1"
+  },
+  "devDependencies": {
+    "gridsome-plugin-remark-shiki": "^0.3.1",
+    "shiki": "^0.1.7",
+    "shiki-languages": "^0.1.6",
+    "unist-util-visit": "^2.0.2"
+  }
+}
