+++ packages/engineer/src/components/icons/IconLogo.vue
