+++ packages/engineer/src/components/icons/IconSearch.vue
