+++ packages/engineer/tailwind.config.js
@@ -49,7 +49,11 @@ module.exports = {
     },
   },
   theme: {
-    extend: {},
+    extend: {
+      colors: {
+        'dark-gray': '#6A737D',
+      },
+    },
   },
   variants: {},
   plugins: [],
