+++ packages/ytailwind/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "ytailwind",
+  "version": "0.0.1",
+  "main": "gridsome.server.js",
+  "dependencies": {
+    "postcss-import": "^12.0.1",
+    "postcss-import-url": "^5.1.0",
+    "postcss-preset-env": "^6.7.0",
+    "tailwindcss": "^1.4.6"
+  }
+}
