--- plugins/shiki/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "yt-shiki",
-  "main": "index.js",
-  "version": "0.0.1"
-}
