--- plugins/ytailwind/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "ytailwind",
-  "version": "0.0.1"
-}
